% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqpcr.R
\name{qqpcr}
\alias{qqpcr}
\title{Title}
\usage{
qqpcr(PCR, reference_gene, control_group, breaks = NULL, custom_colors = NULL)
}
\arguments{
\item{PCR}{PCR_name}

\item{reference_gene}{reference_gene}

\item{control_group}{control_group}

\item{breaks}{breaks}

\item{custom_colors}{custom_colors}
}
\value{
relative expression and plots
}
\description{
Title
}
\examples{
results  <- qqpcr(PCR,'Gapdh',"control")
}
